% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Avg_acc_probs.R
\name{Avg_acc_probs}
\alias{Avg_acc_probs}
\title{Describe the acceptance probabilities.}
\usage{
Avg_acc_probs(res.list, reparam = F)
}
\arguments{
\item{res.list}{The list of results outputted by the MCMC functions.}

\item{reparam}{Whether we are using the reparameterisation to make inference or not (defaults to FALSE).}
}
\value{
A description of the average acceptance probabilities.
}
\description{
This function takes the results from an inference function, and describes in words the acceptance probabilities of the parameters.
 Will show 0% in the case where the parameter is fixed.
}
\examples{

Avg_acc_probs(res.list, reparam = F)


}
\keyword{acceptance}
\keyword{descirption}
\keyword{probabilities}
\keyword{results}
