% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Beta_mat_form.R
\name{Beta_mat_form}
\alias{Beta_mat_form}
\title{Generate an infection rate matrix.}
\usage{
Beta_mat_form(dist.mat, betas, ds)
}
\arguments{
\item{dist.mat}{An NxN matrix of the pairwise distances between individuals.}

\item{betas}{A descending vector of at least length 2 of the different infection rates at each level.}

\item{ds}{An acsending vector of length length(betas)-1 which contains the distances at which the infection rate changes.}
}
\value{
The function returns an NxN matrix of the pairwise infection rates between individuals.
}
\description{
This function generates a beta matrix given a distance function, values of beta, and the distances at which
 the infection rate changes. It works for a heterogeneous population with a near vs. far spatial kernel.
}
\examples{
distance_mat <- Dist_mat_unif(N=100, xlim = 20, ylim = 20)[[2]]
Beta_mat_form(distance_mat, c(0.004, 0.002), 10)
}
\keyword{beta}
\keyword{infection}
\keyword{matrix}
\keyword{rate}
