% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chris_integral_part.R
\name{chris_integral_part}
\alias{chris_integral_part}
\title{Calculate the integral part of the infection likelihood (Chris' code).}
\usage{
chris_integral_part(t_inf_j, events, B)
}
\arguments{
\item{t_inf_j}{A vector of the infection times of all individuals (Inf if not infected), ordered by ID.}

\item{events}{A 2 column matrix where the first column is the infection times, and the second is the paired removal times.}

\item{B}{The infection rate matrix.}
}
\value{
Returns the value of the integral detailed above.
}
\description{
This function is used in the calculation of the likelihood for a General Stochastic Epidemic, and calculates the integral part
 of the infection likelihood, using the alternative double-sum parameterisation.
}
\examples{
This function is utilised by the simple_log_likelihood function.
}
\keyword{GSE}
\keyword{Integral}
\keyword{infection}
\keyword{likelihood}
