% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chris_interval_intersect.R
\name{chris_interval_intersect}
\alias{chris_interval_intersect}
\title{Calculate the interval intersect (Chris' code).}
\usage{
chris_interval_intersect(interval_i, interval_j)
}
\arguments{
\item{interval_i}{Is the matrix cbind(t_inf, t_rem)[infected, ]. This is a 2 column matrix of the infection
and removal times for the infected individuals.}

\item{interval_j}{Is the matrix cbind(t_inf, t_rem). This is a 2 column matrix of the infection
and removal times for all individuals.}
}
\value{
Returns a matrix where each column is a vector for each j.
}
\description{
This function is used in the calculation of the likelihood for a General Stochastic Epidemic,
 and calculates the interval intersect used in the calculation of the intergral for the infection likelihood.
 This is the interval \eqn{[ (R_i ^ I_j) - (I_i ^ I_j) ]} for i in infecteds and j in all.
}
\examples{
This function is utilised by the log_likelihood function.
}
\keyword{GSE}
\keyword{infection}
\keyword{intersection}
\keyword{interval}
\keyword{likelihood}
