% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dist_mat_edge.R
\name{Dist_mat_edge}
\alias{Dist_mat_edge}
\title{Clustered edge population distribution.}
\usage{
Dist_mat_edge(N, xlim = 20, ylim = 20, clusters = 8)
}
\arguments{
\item{N}{The total number of individuals in the population.}

\item{xlim}{The width of the plane on which individuals will be generated (defaults to 20 units wide).}

\item{ylim}{The height of the place on which individuals will be generated (defaults to 20 units high).}

\item{clusters}{The total number of clusters. Half the clusters will be on the right edge, the other half on the left.
(Must be even, greater than 2, and divide N). Clusters are generated on blocks of size (xlim/10, ylim/10).}
}
\value{
The function returns a list. The first object in the list is a Nx2 matrix of the coordinates of the individuals.
  The second object is an NxN distance matrix.
}
\description{
This function generates a distance matrix for a population that is clustered on the left and right edges of the space only.
}
\examples{
xy.coords <- Dist_mat_edge(N=100, xlim = 20, ylim = 20, clusters = 8)[[1]]
distance_mat <- Dist_mat_edge(N=100, xlim = 20, ylim = 20, clusters = 8)[[2]]
plot(xy.coords[,1], xy.coords[,2], type  = "p")

}
\keyword{Clustered}
\keyword{distance}
\keyword{distribution}
\keyword{edge}
\keyword{matrix}
\keyword{population}
