% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dist_mat_latt.R
\name{Dist_mat_latt}
\alias{Dist_mat_latt}
\title{Lattice population distribution (with buddy).}
\usage{
Dist_mat_latt(N, nrow, xlim = 20, ylim = 20, buddy = TRUE)
}
\arguments{
\item{N}{The total number of individuals in the population.}

\item{nrow}{The number of rows of the lattice. N must be divisible by nrow.}

\item{xlim}{The width of the plane on which individuals will be generated (defaults to 20 units wide).}

\item{ylim}{The height of the place on which individuals will be generated (defaults to 20 units high).}

\item{buddy}{Dicatates whether to generate a "buddy" for each point, generated from an isotropic
distribution around each point with variance (xlim/200, ylim/200).}
}
\value{
The function returns a list. The first object in the list is a Nx2 matrix of the coordinates of the individuals.
  The second object is an NxN distance matrix.
}
\description{
This function generates a distance matrix for individuals on a lattice, across a specified rectangular plane. 
    There is also the option to generate a "buddy" for each individual.
}
\examples{
xy.coords <- Dist_mat_latt(N = 200, nrow = 10, xlim = 20, ylim = 20, buddy = TRUE)[[1]]
distance_mat <- Dist_mat_latt(N = 200, nrow = 10, xlim = 20, ylim = 20, buddy = TRUE)[[2]]
plot(xy.coords[,1], xy.coords[,2], type  = "p")

}
\keyword{Lattice}
\keyword{clustered}
\keyword{distance}
\keyword{distribution}
\keyword{matrix}
\keyword{population}
