% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dist_mat_lines.R
\name{Dist_mat_lines}
\alias{Dist_mat_lines}
\title{Spreading lines population distribution.}
\usage{
Dist_mat_lines(N)
}
\arguments{
\item{N}{The total number of individuals in the population.(Must be a square number).}
}
\value{
The function returns a list. The first object in the list is a Nx2 matrix of the coordinates of the individuals.
  The second object is an NxN distance matrix.
}
\description{
This function generates a distance matrix where the population is arranged on a lattice
 where the gaps between rows and columns increases by one unit with each successive individual.
}
\examples{
xy.coords <- Dist_mat_lines(N=100)[[1]]
distance_mat <- Dist_mat_lines(N=100)[[2]]
plot(xy.coords[,1], xy.coords[,2], type  = "p")

}
\keyword{Uniform}
\keyword{distance}
\keyword{distribution}
\keyword{matrix}
\keyword{population}
