% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dist_mat_trials.R
\name{Dist_mat_trials}
\alias{Dist_mat_trials}
\title{Multiple lines population distribution.}
\usage{
Dist_mat_trials(N, nline)
}
\arguments{
\item{N}{The total number of individuals in the population.}

\item{nline}{The number of lines(trials). Must divide N.}
}
\value{
The function returns a list. The first object in the list is a Nx2 matrix of the coordinates of the individuals.
  The second object is an NxN distance matrix.
}
\description{
This function generates a distance matrix where the population is arranged on seperate lines
 where the gaps between rows increase by one unit with each successive individual. The multiple lines are far apart and 
 can be thought of as multiple trials of a 2D epidemic, if the first individual in every row is an initial infective.
}
\examples{
xy.coords <- Dist_mat_trials(N=100, nline = 10)[[1]]
distance_mat <- Dist_mat_trials(N=100, nline = 10)[[2]]
plot(xy.coords[,1], xy.coords[,2], type  = "p")

}
\keyword{Uniform}
\keyword{distance}
\keyword{distribution}
\keyword{matrix}
\keyword{population}
