% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dist_mat_unif.R
\name{Dist_mat_unif}
\alias{Dist_mat_unif}
\title{Uniform population distribution.}
\usage{
Dist_mat_unif(N, xlim = 20, ylim = 20)
}
\arguments{
\item{N}{The total number of individuals in the population.}

\item{xlim}{The width of the plane on which individuals will be generated (defaults to 20 units wide).}

\item{ylim}{The height of the place on which individuals will be generated (defaults to 20 units high).}
}
\value{
The function returns a list. The first object in the list is a Nx2 matrix of the coordinates of the individuals.
  The second object is an NxN distance matrix.
}
\description{
This function generates a distance matrix with uniform distribution of individuals across a specified rectangle.
}
\examples{
xy.coords <- Dist_mat_unif(N=100, xlim = 20, ylim = 20)[[1]]
distance_mat <- Dist_mat_unif(N=100, xlim = 20, ylim = 20)[[2]]
plot(xy.coords[,1], xy.coords[,2], type  = "p")

}
\keyword{Uniform}
\keyword{distance}
\keyword{distribution}
\keyword{matrix}
\keyword{population}
