% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Folded_draw.R
\name{Folded_draw}
\alias{Folded_draw}
\title{Draw a parameter from a folded Normal distribution}
\usage{
Folded_draw(param.cur, param.sigma, lower, upper)
}
\arguments{
\item{param.cur}{The current value of the parameter.}

\item{param.sigma}{The standard deviation of the folded Normal distribution.}

\item{lower}{The lower bound of the distribution.}

\item{upper}{The upper bound of the distribution.}
}
\value{
Returns a draw from the specified folded Normal distribution.
}
\description{
This function draws a single value from a folded Normal distribution with mean the current value of the parameter,
 and variance as required.
}
\examples{
Folded_draw(param.cur = 0.4, param.sigma = 0.1, lower = 0, upper = 1)
}
\keyword{Normal}
\keyword{distribution}
\keyword{folded}
\keyword{proposal}
