% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSE_sim.R
\name{GSE_sim}
\alias{GSE_sim}
\title{An epidemic simulation}
\usage{
GSE_sim(N, beta.mat, gamma)
}
\arguments{
\item{N}{The size of the population.}

\item{beta.mat}{An NxN matrix of the pair wise infection rates between all individuals.}

\item{gamma}{The removal rate.}
}
\value{
The function returns a matrix containing 4 columns; the id number of each individual, their infection time (NA if not infected),
 their removal time (NA if not infected), and their generated infectious period length.
}
\description{
This function allows you to simulate a stochastic SIR epidemic, with a heterogeneous contact matrix.
}
\details{
The infectious life history of each individual is comprised of two parts;
\itemize{
 \item \eqn{Q_i}, the length of individuals \eqn{i}'s infectious period.
 \item \eqn{W_{i,j}}, the length of time, after individual \eqn{i} becomes infected, before they make contact with individual \eqn{j}.
}
The distributions of which are given by;
\itemize{
 \item \eqn{Q_i \overset{iid}{\sim}}{Q_i ~} Exp(\eqn{\gamma}{gamma}).
 \item \eqn{W_{i,j} \overset{iid}{\sim}}{W_{i,j} ~} Exp(\eqn{\beta_{i,j}}{beta_{i,j}}).
}
The simulation begins with an initial infective infectious at time 0.

\eqn{I_i} is the time at which individual \eqn{i} becomes 
infected, and their removal time is given by \eqn{R_i = I_i + Q_i}.

If \eqn{W_{i,j} < Q_i}, then individual \eqn{i} makes infectious contact with individual \eqn{j} at time \eqn{I_i + W_{i,j}}.

If individual \eqn{j} is susceptible when this happens, then they become infected, otherwise nothing happens. 
The simulation assumes that \eqn{W_{i,j} != W_{j,i}}.

The simulation continues until there are no infected individuals left in the population.
}
\examples{

# Generate a distance matrix
   distance_mat <- Dist_mat_unif(N=100, xlim = 20, ylim = 20)[[2]]

# Generate an associated infection rate rate matrix
   rate_mat <- Beta_mat_form(distance_mat, c(0.004, 0.002), 10)
   
# Generate a simulated epidemic
   Hetero_sim <- GSE_sim(N = 100, beta.mat = rate_mat, gamma = 0.15)
}
\keyword{Epidemic}
\keyword{GSE}
\keyword{General}
\keyword{Stochastic}
