% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hist_posteriors.R
\name{Hist_posteriors}
\alias{Hist_posteriors}
\title{Plot the trace plots of the accepted samples.}
\usage{
Hist_posteriors(res, params, burnin = NA)
}
\arguments{
\item{res}{The results matrix from the MCMC functions.}

\item{params}{A named vector of the parameters, with names "beta1", "beta2", "gamma", "d", "p", and "llh".}

\item{burnin}{The ammount of burnin that wants to be removed (defaults to NA to plot the full trace plots).}
}
\value{
A description of the average acceptance probabilities.
}
\description{
This function takes the results from an inference function, and plots the trace plot of all the accepted samples,
 with a red line to represent the true value if it falls within the limits.
 If a parameter was fixed it will plot a straight line.
}
\examples{

Plot_trace(res, params, burnin = NA)


}
\keyword{acceptance}
\keyword{description}
\keyword{probabilities}
\keyword{results}
