% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Initialise_2b.R
\name{Initialise_2b}
\alias{Initialise_2b}
\title{Initialise the parameters and infection times.}
\usage{
Initialise_2b(N = N, rem.times = rem.times, beta1.true = NA,
  beta2.true = NA, p.true = NA, dist.true = NA, inf.times = NA,
  d.lower = d.lower, d.upper = d.upper, dist.mat, reparam = FALSE)
}
\arguments{
\item{N}{The size of the total population.}

\item{rem.times}{A vector of the removal times.}

\item{beta1.true}{The true value of beta1, will fix beta1 if given, defaults to NA.}

\item{beta2.true}{The true value of beta2, will fix beta2 if given, defaults to NA.}

\item{p.true}{The true value of p, will fix p if given, defaults to NA.}

\item{dist.true}{The true value of d, will fix d if given, defaults to NA.}

\item{inf.times}{A vector of the infection times.}

\item{d.lower}{Lower bound on the proposal of d.}

\item{d.upper}{Upper bound on the proposal of d.}

\item{dist.mat}{An NxN distance matrix.}

\item{reparam}{TRUE/FALSE statement of whether we are using the reparameterisation beta2 = p * beta1 (p in (0,1)).}
}
\value{
Returns a list with values (beta1.cur, beta2.cur, p.cur, dist.cur, inf.times).
}
\description{
This function initialises an epidemic (checking it is valid) which has two infection rates.
}
\examples{
Initilise_2b(beta1.true = 0.004, dist.mat = distance.mat, reparam = T)
}
\keyword{Initialise}
\keyword{betas}
\keyword{epidemic}
\keyword{two}
