% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integral_part.R
\name{integral_part}
\alias{integral_part}
\title{Calculate the integral part of the infection likelihood.}
\usage{
integral_part(inf_times, rem_times, B, infected_inds)
}
\arguments{
\item{inf_times}{A vector of the infection times of all individuals (Inf if not infected).}

\item{rem_times}{A vector of the removal times of all infected individuals.}

\item{B}{The infection rate matrix.}

\item{infected_inds}{A vector of the infected individuals.}
}
\value{
Returns the value of the integral detailed above.
}
\description{
This function is used in the calculation of the likelihood for a General Stochastic Epidemic, and calculates the integral part
 of the infection likelihood, using the alternative double-sum parameterisation.
}
\examples{
This function is utilised by the log_likelihood function.
}
\keyword{GSE}
\keyword{Integral}
\keyword{infection}
\keyword{likelihood}
