% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval_intersect.R
\name{interval_intersect}
\alias{interval_intersect}
\title{Calculate the interval intersect.}
\usage{
interval_intersect(inf_times, rem_times, infected_inds)
}
\arguments{
\item{inf_times}{A vector of the infection times of all individuals (Inf if not infected).}

\item{rem_times}{A vector of the removal times of all infected individuals.}

\item{infected_inds}{A vector of the infected individuals.}
}
\value{
Returns a matrix where each column is a vector for each j.
}
\description{
This function is used in the calculation of the likelihood for a General Stochastic Epidemic, and calculates the interval
 intersect used in the calculation of the intergral for the infection likelihood.
 This is the interval \eqn{[ min(R_i, I_j) - min(I_i, I_j) ]} for i in infecteds, j in all.
}
\examples{
This function is utilised by the log_likelihood function.
}
\keyword{GSE}
\keyword{infection}
\keyword{intersection}
\keyword{interval}
\keyword{likelihood}
