% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_likelihood.R
\name{log_likelihood}
\alias{log_likelihood}
\title{Calculate the log-likelihood for an epidemic, up to a constant of proportionality.}
\usage{
log_likelihood(inf_times, rem_times, B)
}
\arguments{
\item{inf_times}{A vector of the infection times of all individuals (Inf if not infected).}

\item{rem_times}{A vector of the removal times of all infected individuals.}

\item{B}{The infection rate matrix.}
}
\value{
Returns the value of the log likelihood, up to a constant of proportionality.
}
\description{
This function is used tp calculate the likelihood of a General Stochastic Epidemic, and brings together all the previously 
 specified functions; prod_part, interval_intersect, and integral_part.
}
\examples{
This function is utilised in the inference functions.
}
\keyword{GSE}
\keyword{infection}
\keyword{likelihood}
\keyword{log}
\keyword{loglikelihood}
