% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MH_accept.R
\name{MH_accept}
\alias{MH_accept}
\title{Calculate the acceptance probability of a Metropolis-Hastings step (ratio of likelihoods case only.)}
\usage{
MH_accept(llh, llh.prime)
}
\arguments{
\item{llh}{The value of the log-likelihood with the current value of the paramter in question.}

\item{llh.prime}{The value of the log-likelihood with the newly drawn value of the paramter in question.}
}
\value{
Returns the min(1, ratio of likelihoods).
}
\description{
This function calculates alpha, the acceptance probability of a MH step, for a random walk or similar where the proposal is
 symmetric so cancels and one is left with a ratio of likelihoods.
}
\examples{
MH_accept(llh, llh.prime)
}
\keyword{Hastings}
\keyword{MH}
\keyword{Metropolis}
\keyword{Metropolis-Hastings}
\keyword{acceptance}
\keyword{alpha}
\keyword{probability}
