% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MH_accept_MRW.R
\name{MH_accept_MRW}
\alias{MH_accept_MRW}
\title{Calculate the acceptance probability of a Metropolis-Hastings step for a multiplicative random walk.}
\usage{
MH_accept_MRW(llh, llh.prime, nu.prior, lambda.prior, param.cur,
  param.prime)
}
\arguments{
\item{llh}{The value of the log-likelihood with the current value of the paramter in question.}

\item{llh.prime}{The value of the log-likelihood with the newly drawn value of the paramter in question.}

\item{nu.prior}{Assuming a Gamma prior on the parameter in question, the value of the parameter nu in that prior.}

\item{lambda.prior}{Assuming a Gamma prior on the parameter in question, the value of the parameter lambda in that prior.}

\item{param.cur}{The current value of the parameter of interest.}

\item{param.prime}{The proposed value of the parameter of interest.}
}
\value{
Returns the min(1, alpha).
}
\description{
This function calculates alpha, the acceptance probability of a MH step, for a multiplicative random walk, which is where proposals
 are made on the log scale. It assumes a Gamma prior on the parameter of interest.
}
\examples{
MH_accept_MRW(llh, llh.prime, nu.beta, lambda.beta, beta.cur, beta.prime)
}
\keyword{Hastings}
\keyword{MH}
\keyword{Metropolis}
\keyword{Metropolis-Hastings}
\keyword{acceptance}
\keyword{alpha}
\keyword{multiplicative}
\keyword{probability}
\keyword{random}
\keyword{walk}
