% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/P_mat_form.R
\name{P_mat_form}
\alias{P_mat_form}
\title{Generate an proportion of maximum infection rate matrix.}
\usage{
P_mat_form(dist.mat, ps = 1, ds)
}
\arguments{
\item{dist.mat}{An NxN matrix of the pairwise distances between individuals.}

\item{ps}{A descending vector of at least length 2 of the different proportions of maximum infection rate at each level.
The first element is assumed to be 1.}

\item{ds}{An acsending vector of length length(betas)-1 which contains the distances at which the infection rate changes.}
}
\value{
The function returns an NxN matrix of the pairwise proportion of maximum infection rates between individuals.
}
\description{
This function generates a p matrix given a distance function, values of beta, and the distances at which
 the infection rate changes. It works for a heterogeneous population with a near vs. far spatial kernel.
 It is essentially just a wrapper for the Beta_mat_form function, but used to calculate p.
}
\examples{
distance_mat <- Dist_mat_unif(N=100, xlim = 20, ylim = 20)[[2]]
P_mat_form(distance_mat, c(1, 0.1), 10)
}
\keyword{infection}
\keyword{matrix}
\keyword{p}
\keyword{proportion}
\keyword{rate}
