% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_epidemic.R
\name{Plot_epidemic}
\alias{Plot_epidemic}
\title{Plot a spatial epidemic.}
\usage{
Plot_epidemic(xy.coords, sim, init.inf = 1)
}
\arguments{
\item{xy.coords}{The coordinates of all individuals in the population}

\item{sim}{The output of the epidemic simulation}

\item{init.inf}{A vector of the ID(s) of the initial invective(s). This just defaults to 1 as in the simulation.}
}
\value{
A spatial plot of the epidemic
}
\description{
This function takes in an epidemic simulation generated by this package and outputs a colour coded plot.
 The circles are the remaining susceptibles, the red crosses are the infected individuals, and the blue crossed boxes are the
 initial infectives
}
\examples{
# Generate a distance matrix
   C <- Dist_mat_unif(N=100, xlim = 20, ylim = 20)
   xy.coords <- C[[1]]
   distance_mat <- C[[2]]

# Generate an associated infection rate rate matrix
   rate_mat <- Beta_mat_form(distance_mat, c(0.004, 0.002), 10)
   
# Generate a simulated epidemic
   Hetero_sim <- HeteroGSEsim(N = 100, beta.mat = rate_mat, gamma = 0.15)
   
# Plot the epidemic
   Plot_epidemic(xy.coords, Hetero_sim, init.inf = c(1))
}
\keyword{colour}
\keyword{epidemic}
\keyword{plot}
