% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prod_part.R
\name{prod_part}
\alias{prod_part}
\title{Calculate the product part of the infection likelihood}
\usage{
prod_part(inf_times, rem_times, B, infected_inds)
}
\arguments{
\item{inf_times}{A vector of the infection times of all individuals (Inf if not infected).}

\item{rem_times}{A vector of the removal times of all infected individuals.}

\item{B}{The infection rate matrix.}

\item{infected_inds}{A vector of the infected individuals.}
}
\value{
Returns the value of the product detailed above.
}
\description{
This function is used in the calculation of the likelihood for a General Stochastic Epidemic, and calculates 
 \eqn{\prod_{j != kappa}^{n_I} [ \sum_{ i in I_{n_{j-}} } [ beta_{i,j} ] ]}.
}
\examples{
This function is utilised by the log_likelihood function.
}
\keyword{GSE}
\keyword{Product}
\keyword{infection}
\keyword{likelihood}
