% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seed_finder.R
\name{Seed_finder}
\alias{Seed_finder}
\title{Find a seed for the desired outbreak size}
\usage{
Seed_finder(model, ..., betas, ds, size, error, seeds)
}
\arguments{
\item{model}{Takes values "Unif", "Clust", "Edge", "Latt", "Lines", or "Trials", corresponding to the appropriate distance matrix function.}

\item{...}{The parameters of the specifief distance matrix function (see example).}

\item{betas}{A vector of infection rates, used for the Beta_mat_form function.}

\item{ds}{A vector of distances at which the infection rate changes, used for the Beta_mat_form function.}

\item{size}{The final size of the epidemic that is desired (total number infected).}

\item{error}{The allowed error on the final size above and below.}

\item{seeds}{The number of seeds desired to be tested. Always iterates from 1:seeds.}
}
\value{
A histogram of the final size of all tested simulations, and a table of seeds and final sizes that fit the requirements.
}
\description{
Runs a series of independent simulations for a series of seeds, and outputs a seed(s) that give a desired final size,
 as well as a histogram of all final sizes.
}
\examples{
Seed_finder("Clust", 
            N=100, xlim = 20, ylim = 20, centres = 5, spread = 1, 
            betas = c(0.1, 0.001), ds = 0.25, 
            size = 25, error = 2, seeds = 250)
}
\keyword{Histogram}
\keyword{final}
\keyword{seed}
\keyword{seeds}
\keyword{size}
