% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_integral_part.R
\name{simple_integral_part}
\alias{simple_integral_part}
\title{Calculate the integral part of the infection likelihood (Simple code).}
\usage{
simple_integral_part(I = inf.times, R = rem.times,
  beta.mat = beta.mat)
}
\arguments{
\item{I}{A vector of the infection times of all individuals (Inf if not infected), ordered by ID.}

\item{R}{A vector of the infection times of all individuals (Inf if not infected), ordered by ID.}

\item{beta.mat}{The infection rate matrix.}
}
\value{
Returns the value of the integral detailed above.
}
\description{
This function is used in the calculation of the likelihood for a General Stochastic Epidemic, and calculates the integral part
 of the infection likelihood, using the alternative double-sum parameterisation.
}
\examples{
This function is utilised by the simple_log_likelihood function.
}
\keyword{GSE}
\keyword{Integral}
\keyword{infection}
\keyword{likelihood}
