% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_log_likelihood.R
\name{simple_log_likelihood}
\alias{simple_log_likelihood}
\title{Calculate the log-likelihood for an epidemic, up to a constant of proportionality (Simple code).}
\usage{
simple_log_likelihood(inf.times, rem.times, beta.mat)
}
\arguments{
\item{inf.times}{A vector of the infection times of all individuals (Inf if not infected).}

\item{rem.times}{A vector of the removal times of all infected individuals.}

\item{beta.mat}{The infection rate matrix.}
}
\value{
Returns the value of the log likelihood, up to a constant of proportionality.
}
\description{
This function is used tp calculate the likelihood of a General Stochastic Epidemic, and brings together all the previously 
 specified functions; prod_part, interval_intersect, and integral_part.
}
\examples{
This function is utilised in the inference functions.
}
\keyword{GSE}
\keyword{infection}
\keyword{likelihood}
\keyword{log}
\keyword{loglikelihood}
