% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_prod_part.R
\name{simple_prod_part}
\alias{simple_prod_part}
\title{Calculate the product part of the infection likelihood (Simple code)}
\usage{
simple_prod_part(inf.times, rem.times, beta.mat)
}
\arguments{
\item{inf.times}{A vector of the infection times of all individuals (Inf if not infected), ordered by ID.}

\item{rem.times}{A vector of the removal times of all individuals (Inf if not infected), ordered by ID.}

\item{beta.mat}{The infection rate matrix.}
}
\value{
Returns the value of the product detailed above.
}
\description{
This function is used in the calculation of the likelihood for a General Stochastic Epidemic, and calculates
 \eqn{\prod_{j != kappa}^{n_I} [ \sum_{ i in I_{n_{j-}} } [ beta_{i,j} ] ]}.
}
\examples{
This function is utilised by the simple_log_likelihood function.
}
\keyword{GSE}
\keyword{Product}
\keyword{infection}
\keyword{likelihood}
