% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval_intersect.R
\name{interval_intersect}
\alias{interval_intersect}
\title{Calculate the interval intersect.}
\usage{
interval_intersect(inf_times, rem_times, infected_inds)
}
\arguments{
\item{interval_i}{Is the matrix cbind(t_inf, t_rem)[infected, ]. This is a 2 column matrix of the infection ]
and removal times for the infected individuals.}

\item{interval_j}{Is the matrix cbind(0, t_inf). This is a 2 column matrix of 0s in the first column and
the infection times in the second.}
}
\value{
Returns a matrix where each column is a vector for each j.
}
\description{
This function is used in the calculation of the likelihood for a General Stochastic Epidemic, and calculates the interval
 intersect used in the calculation of the intergral for the infection likelihood. 
 This is the interval \eqn{[ (R_i ^ I_j) - (I_i) ^ (I_j) ]} for all i and j.
}
\examples{
This function is utilised by the log_likelihood function.
}
\keyword{GSE}
\keyword{infection}
\keyword{intersection}
\keyword{interval}
\keyword{likelihood}
